backend/requirements.txt
fastapi
uvicorn
sqlalchem